requirements.txt
requests>=2.25.0
eth-account>=0.8.0
pycryptodome>=3.15.0 